Audio: begin cleanup
